Europeana Search Api

The Search API allows to search the Europeana repository for metadata records and media. The Search API is the most straightforward to use. It works in a similar fashion to the Europeana website when it comes to interacting with the data. You can use the API to search for keywords, and it will return any entries that contain those words. You can refine your search with more advanced queries like Boolean Searches, or you can filter out parts of the results advanced filtering.

Usage

query_search_api(
  query = NULL,
  rows = NULL,
  profile = NULL,
  qf = NULL,
  reusability = NULL,
  media = NULL,
  thumbnail = NULL,
  landingpage = NULL,
  colourpalette = NULL,
  facet = NULL,
  limit = NULL,
  start = NULL,
  sort = NULL,
  path = "/record/v2/search.json",
  ...
)

Source

https://pro.europeana.eu/page/search

Arguments

query

(character) string with the search term(s)

rows

(numeric) that indicates the number of records to return

profile

(character) Profile parameter controls the format and richness of the response.

qf

(character) Facet filtering query. This parameter can be defined more than once.

reusability

(character) Filter by copyright status.

media

(character) Filter by records where an URL to the full media file is present in the edm:isShownBy or edm:hasView metadata and is resolvable.

thumbnail

(character) Filter by records where a thumbnail image has been generated for any of the WebResource media resources (thumbnail available in the edmPreview field).

landingpage

(character) Filter by records where the link to the original object on the providers website (edm:isShownAt) is present and verified to be working.

colourpalette

(character) Filter by images where one of the colours of an image matches the provided colour code. You can provide this parameter multiple times, the search will then do an 'AND' search on all the provided colours. See colour palette.

facet

(character) Name of an individual facet. See individual facets.

limit

(numeric) The number of records to return. Maximum is 100. Default: 10

start

(numeric) The item in the search results to start with. The first item is 1. Default: 1

sort

(character) Sort by a field, e.g., timestamp_update+desc

path

(character) URL signature with the API version

...

parameters passed in get request

Details

In the `query` parameter the Apache Lucene Query Syntax is inheritly supported by the Search API. Users can use Lucene and Apache SOLR guides to get the most out of the Europeana repository.

The response is always formatted in JSON and will contain a number of fields that present information about the handling of the request, while the concrete information about the record is presented in the "items" field (see Metadata Sets). The parsed information is stored in the `content` field of the S3 object returned.

References

Doerr M, Gradmann S, Hennicke S, Isaac A, Meghini C, Van de Sompel H (2010). “The europeana data model (edm).” In World Library and Information Congress: 76th IFLA general conference and assembly, volume 10, 15.

Wickham H (2020). httr: Tools for Working with URLs and HTTP. https://httr.r-lib.org/, https://github.com/r-lib/httr.

Ooms J (2014). “The jsonlite Package: A Practical and Consistent Mapping Between JSON Data and R Objects.” arXiv:1403.2805 [stat.CO]. https://arxiv.org/abs/1403.2805.

Examples

if (FALSE) { # Sys.getenv("EUROPEANA_KEY") != ""
# \donttest{
#set your API key with set_key(api_key = "XXXX")
#query search API
res <- query_search_api("arioch", qf = "1712", media = TRUE)
# }
}